+++ src/Home.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+
+class Home extends Component {
+    render() {
+        return (
+
+            <div>
+
+                <img src="https://waylostreams.live/waylostreams.jpg"></img>
+                <h2>This is Waylostreams</h2>
+            </div>
+
+
+        );
+    }
+}
+
+export default Home;
